// src/app/model/DataDemo.ts
export class DataDemo{
    shape: string;
    Numbering: string;
    reactionDate: string;
    dormitoryStaff:string;
    dormitoryEast:string;
    location :string;
    abnormal: string;
    abnormalName: string;
    description: string;
    telephoneContact: string;
    managerStaff: string;
    dateofacceptance:string;
    Fillintheday: string;
    Actualday:string;

}